+++ src/main/java/moduleseven/VideoFile.java
@@ -6,7 +6,7 @@ package moduleseven;
 public class VideoFile extends File {
     protected static final String VIDEO = ".avi" ;
 
-    public VideoFile(String location, String name, String extension, float size) {
+    public VideoFile(String location, String name, String extension, int size) {
         super(location, name, extension, size);
     }
